chore: Move translations again to please CrowdIn
