+++ crowdin.yml
@@ -1,3 +1,3 @@
 files:
-  - source: /shared/translations/locales/en_US.json
-    translation: /shared/translations/locales/%locale_with_underscore%.json
+  - source: /shared/i18n/locales/en_US.json
+    translation: /shared/i18n/locales/%locale_with_underscore%/translation.json
