+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "clean": "rimraf build",
-    "build:i18n": "i18next 'app/**/*.js' 'server/**/*.js' && cp -R ./shared/translations/locales/*.json ./build/shared/translations/locales",
+    "build:i18n": "i18next 'app/**/*.js' 'server/**/*.js' && cp -R ./shared/i18n/locales ./build/shared/i18n/locales",
     "build:server": "babel -d ./build/server ./server && babel -d ./build/shared ./shared && cp package.json ./build && ln -sf \"$(pwd)/webpack.config.dev.js\" ./build",
     "build:webpack": "webpack --config webpack.config.prod.js",
     "build": "yarn clean && yarn build:webpack && yarn build:i18n && yarn build:server",
